B3.  Include features that allow the user to do the following for each vacation:
a.  Display a detailed view of the vacation, including all vacation details. This view can also be used to add and update the vacation information.

b.  Enter, edit, and delete vacation information.

c.  Include validation that the input dates are formatted correctly.

d.  Include validation that the vacation end date is after the start date.

e.  Include an alert that the user can set which will trigger on the start and end date, displaying the vacation title and whether it is starting or ending.

f.  Include sharing features so the user can share all the vacation details via a sharing feature (either e-mail, clipboard or SMS) that automatically populates with the vacation details.

g.  Display a list of excursions associated with each vacation.

h.  Add, update, and delete as many excursions as needed.