b5.  Include features that allow the user to do the following for each excursion:
a.  Display a detailed view of the excursion, including the title, and date

b.  Enter, edit, and delete excursion information.

c.  Include validation that the input dates are formatted correctly.

d.  Include an alert that the user can set that will trigger on the excursion date, stating the excursion title.

e.  Include validation that the excursion date is during the associated vacation.